feat: refactor skeleton component
